Rename: 테스트하기 쉽도록 non-lazy 함수 분리
