+++ src/4week/utils/non-lazy/unique.js
@@ -1,0 +1,10 @@
+const { I } = require('../../internal_utils');
+
+const unique = (arr) => {
+  if (!I.isIterable(arr)) {
+    return [];
+  }
+  return [...new Set(arr)];
+}
+
+exports.unique = unique;
